requirements.txt
cassandra-csv==0.3
cassandra-driver==3.25.0
pyspark==2.4.5
